refactor: extract subagent artifact poller
